Made some progress
The navbar category pages are now created! If a users clicks on a category, the user will navigate to the news page of that chosen category.
On this news page I made a searchbar (dynamical form). 
If the user types something in the input field, the searchbar will show the available countries to navigate to.
Here the user can read news about the chosen category of that country.

My plan for the next time is to make it possible for the user to navigate to a specific country on the chosen category page.
So lets say someone is a fan of American politics they just have to navigate to the politic category on the navbar and there the user can type in the United states of america to see some news about american politics.

Biggest Updates:
App.js;
News.js;
Navbar.js;
searchbar.js;
DropdownOptionsNavbar.js;